requirements-dev.txt
codacy-coverage
pytest<4.1  # see https://github.com/pytest-dev/pytest-cov/issues/252
pytest-cov>=2.6.1,<3
flake8
ipdb==0.10.3
